# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Elegant Sip API",
    "version": "0.1.0",
    "description": "Backend for the Elegant Sip storefront — single-origin Darjeeling tea.\n\n**Money is integer paise** (1 rupee = 100 paise). Razorpay settles in paise and\nfloats cannot represent money. Totals are always a whole number of rupees:\nGST is rounded with the same rule the storefront uses, so the figure a customer\nis shown and the figure they are charged cannot drift apart.\n\nErrors share one shape across every endpoint, modelled on RFC 9457. `detail` is\nalways safe to display; `requestId` ties a response to the server logs.",
    "contact": {
      "name": "Elegant Sip",
      "email": "elegantsipdarjeeling@gmail.com"
    }
  },
...
  "paths": {
...
    "/v1/admin/products": {
      "post": {
        "summary": "Add a product",
        "tags": [
          "Admin"
        ],
        "description": "Creates a product and its variants in one transaction.\n\nA tea with no price yet should be created as `coming-soon` with a zero-price,\nzero-stock variant. The storefront then shows a badge rather than \"₹0\", which\nwould read as free, and no JSON-LD offer is emitted for it.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "slug": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                    "description": "URL-safe identifier, e.g. \"first-flush-whole-leaf\""
                  },
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "status": {
                    "default": "active",
                    "type": "string",
                    "enum": [
                      "active",
                      "coming-soon"
                    ]
                  },
                  "description": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 600
                  },
                  "longDescription": {
                    "type": "string",
                    "maxLength": 4000
                  },
                  "imageSrc": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "category": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 60
                  },
                  "tastingNotes": {
                    "default": [],
                    "maxItems": 8,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "maxLength": 40
                    }
                  },
                  "bodyLevel": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "harvestLabel": {
                    "type": "string",
                    "maxLength": 60
                  },
                  "origin": {
                    "type": "object",
                    "properties": {
                      "origin": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 120
                      },
                      "estate": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 80
                      },
                      "elevation": {
                        "type": "string",
                        "maxLength": 60
                      },
                      "harvest": {
                        "type": "string",
                        "maxLength": 60
                      },
                      "cultivar": {
                        "type": "string",
                        "maxLength": 120
                      }
                    },
                    "required": [
                      "origin",
                      "estate",
                      "elevation",
                      "harvest",
                      "cultivar"
                    ]
                  },
                  "flavorProfile": {
                    "type": "object",
                    "properties": {
                      "strength": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 5
                      },
                      "astringency": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 5
                      },
                      "sweetness": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 5
                      },
                      "floral": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 5
                      },
                      "caffeine": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 5
                      }
                    },
                    "required": [
                      "strength",
                      "astringency",
                      "sweetness",
                      "floral",
                      "caffeine"
                    ]
                  },
                  "brewingGuide": {
                    "type": "object",
                    "properties": {
                      "temperature": {
                        "type": "string",
                        "maxLength": 40
                      },
                      "time": {
                        "type": "string",
                        "maxLength": 40
                      },
                      "steeps": {
                        "type": "string",
                        "maxLength": 20
                      },
                      "leafAmount": {
                        "type": "string",
                        "maxLength": 40
                      },
                      "notes": {
                        "type": "string",
                        "maxLength": 400
                      }
                    },
                    "required": [
                      "temperature",
                      "time",
                      "steeps",
                      "leafAmount",
                      "notes"
                    ]
                  },
                  "gardenSlug": {
                    "description": "Links the product to a garden, in both directions",
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80,
                    "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
                  },
                  "variants": {
                    "minItems": 1,
                    "maxItems": 10,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "size": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 60,
                          "description": "e.g. \"Premium · 100 g\""
                        },
                        "price": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991,
                          "description": "Amount in paise (1 rupee = 100 paise)"
                        },
                        "stock": {
                          "default": 0,
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "size",
                        "price"
                      ]
                    }
                  }
                },
                "required": [
                  "slug",
                  "name",
                  "description",
                  "imageSrc",
                  "category",
                  "variants"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "slug": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 80,
                      "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                      "description": "URL-safe identifier, e.g. \"first-flush-whole-leaf\""
                    },
                    "variantCount": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "slug",
                    "variantCount"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "Stable machine-readable slug, e.g. \"validation_error\", \"out_of_stock\""
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string",
                      "description": "What went wrong and what to do about it — displayable as-is"
                    },
                    "errors": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Correlates this response with the server logs"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "Stable machine-readable slug, e.g. \"validation_error\", \"out_of_stock\""
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string",
                      "description": "What went wrong and what to do about it — displayable as-is"
                    },
                    "errors": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Correlates this response with the server logs"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/admin/products/{slug}/variants": {
      "post": {
        "summary": "Add a variant",
        "tags": [
          "Admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "size": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 60
                  },
                  "price": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991,
                    "description": "Amount in paise (1 rupee = 100 paise)"
                  },
                  "stock": {
                    "default": 0,
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  }
                },
                "required": [
                  "size",
                  "price"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            },
            "in": "path",
            "name": "slug",
            "required": true,
            "description": "URL-safe identifier, e.g. \"first-flush-whole-leaf\""
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "slug": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 80,
                      "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                      "description": "URL-safe identifier, e.g. \"first-flush-whole-leaf\""
                    },
                    "size": {
                      "type": "string"
                    },
                    "price": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991,
                      "description": "Amount in paise (1 rupee = 100 paise)"
                    },
                    "stock": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    }
                  },
                  "required": [
                    "slug",
                    "size",
                    "price",
                    "stock"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "Stable machine-readable slug, e.g. \"validation_error\", \"out_of_stock\""
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string",
                      "description": "What went wrong and what to do about it — displayable as-is"
                    },
                    "errors": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Correlates this response with the server logs"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "Stable machine-readable slug, e.g. \"validation_error\", \"out_of_stock\""
                    },
                    "title": {
                      "type": "string",
                      "description": "Short human-readable summary"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string",
                      "description": "What went wrong and what to do about it — displayable as-is"
                    },
                    "errors": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Correlates this response with the server logs"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "requestId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    },
    {
      "url": "https://api.elegantsip.com",
      "description": "Production"
    }
  ],
...
}